sync

+++ Bolt-Core/include/Graphic/Texture/Texture.hpp
@@ -17,8 +17,8 @@ namespace bolt {
 	class Texture {
 	protected:
 		u32 m_id = 0;
-		u16 m_width = 0;
-		u16 m_height = 0;
+		u32 m_width = 0;
+		u32 m_height = 0;
 		texture::Parameters m_parameters{};
 
 		b8 m_generated = false;
@@ -28,7 +28,7 @@ namespace bolt {
 	public:
 		Texture() = default;
 
-		Texture(const texture::Parameters& params, const u16 width, const u16 height) : m_width(width), m_height(height), m_parameters(params) { this->m_ready = true; }
+		Texture(const texture::Parameters& params, const u32 width, const u32 height) : m_width(width), m_height(height), m_parameters(params) { this->m_ready = true; }
 
 		~Texture() = default;
 
@@ -56,8 +56,8 @@ namespace bolt {
 
 		b8 isTextureCreated() const { return this->m_created; }
 
-		inline u16 getWidth() const { return this->m_width; }
+		inline u32 getWidth() const { return this->m_width; }
 
-		inline u16 getHeight() const { return this->m_height; }
+		inline u32 getHeight() const { return this->m_height; }
 	};
 } // namespace Bolt
